requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argcomplete==3.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
black==23.7.0
blas==1.0
blinker==1.8.2
braceexpand==0.1.7
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2024.7.2
cachetools==5.3.3
cd-fvd==0.1.1
certifi==2024.7.4
chardet==5.1.0
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
cmake==3.14.4
colorama==0.4.6
comm==0.2.2
configargparse==1.7
contourpy==1.2.1
cuda-cudart==11.7.99
cuda-cupti==11.7.101
cuda-libraries==11.7.1
cuda-nvrtc==11.7.99
cuda-nvtx==11.7.91
cuda-runtime==11.7.1
cuda-version==12.5
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.10.1
diff-gauss==1.0.10.0
dnspython==2.6.1
docker-pycreds==0.4.0
dust3r==0.1.0
einops==0.8.0
email-validator==2.2.0
evo==1.29.0
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.13
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpeg==4.3
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
flask==3.0.3
fonttools==4.53.0
freetype==2.12.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gdown==5.2.0
gitdb==4.0.11
gitpython==3.1.43
gmp==6.2.1
gmpy2==2.1.2
gnutls==3.6.15
gradio-client==1.0.2
grpcio==1.64.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==8.0.0
importlib-resources==6.4.0
intel-openmp==2023.1.0
invisible-watermark==0.2.0
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jpeg==9e
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
kornia==0.6.9
lame==3.100
lcms2==2.12
ld_impl_linux-64==2.38
lerc==3.0
libcublas==11.10.3.66
libcufft==10.7.2.124
libcufile==1.10.0.4
libcurand==10.3.6.39
libcusolver==11.4.0.1
libcusparse==11.7.4.91
libdeflate==1.17
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libiconv==1.16
libidn2==2.3.4
libnpp==11.7.4.75
libnvjpeg==11.8.0.2
libpng==1.6.39
libstdcxx-ng==11.2.0
libtasn1==4.19.0
libtiff==4.5.1
libunistring==0.9.10
libwebp-base==1.3.2
lightning-utilities==0.11.3.post0
lit==18.1.8
lz4==4.3.3
lz4-c==1.9.4
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
mpc==1.1.0
mpfr==4.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
nbformat==5.10.4
ncurses==6.4
nest-asyncio==1.6.0
nettle==3.7.3
networkx==3.2.1
ninja==1.11.1.1
numexpr==2.10.1
numpy==1.26.4
numpy-base==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open-clip-torch==2.24.0
open3d==0.18.0
opencv-python==4.6.0.66
openh264==2.1.1
openjpeg==2.4.0
openssl==3.0.14
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
plyfile==1.0.3
portalocker==2.10.1
prompt-toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pudb==2024.1
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==1.10.17
pydantic-core==2.20.0
pydeck==0.9.1
pydub==0.25.1
pyglet==1.5.29
pygments==2.18.0
pyparsing==3.1.2
pyquaternion==0.9.9
pysocks==1.7.1
python==3.9.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-cuda==11.7
pytorch-fid==0.3.0
pytorch-lightning==2.0.1
pytorch-mutex==1.0
pytz==2024.1
pywavelets==1.6.0
pyyaml==6.0.1
readline==8.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
retrying==1.3.4
rich==13.7.1
roma==1.5.0
rosbags==0.9.23
rpds-py==0.18.1
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
ruff==0.5.0
safetensors==0.4.3
sam-2==1.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==69.5.1
shellingham==1.5.4
simple-knn==0.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlite==3.45.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
sympy==1.12
tbb==2021.8.0
tenacity==8.4.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardx==2.6
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.7
tk==8.6.14
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.4.0.post0
torchtriton==2.0.0
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.19.1
trimesh==4.4.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
dreamdrive==0.0.0
urllib3==1.26.19
urwid==2.6.14
urwid-readline==0.14
uvicorn==0.30.1
uvloop==0.19.0
vwm==0.1.0
wandb==0.17.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
xformers==0.0.22
xz==5.4.6
yarl==1.9.4
zipp==3.19.2
zlib==1.2.13
zstandard==0.22.0
zstd==1.5.5
